Refactored TaskFactory into its own module. Cleaned up types and documentation.
